+++ app/Controllers/GoogleAnalyticsController.php
@@ -150,7 +150,7 @@ class GoogleAnalyticsController extends Controller
             exit;
         }
 
-        $optionsJson = json_encode($result['properties'], JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES);
+        $optionsJson = json_encode($result['properties'], JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES | JSON_HEX_TAG | JSON_HEX_AMP | JSON_HEX_APOS | JSON_HEX_QUOT);
 
         $body = <<<'HTML'
         <div class="p-card">