+++ app/Controllers/SearchConsoleController.php
@@ -153,7 +153,7 @@ class SearchConsoleController extends Controller
             exit;
         }
 
-        $optionsJson = json_encode($sitesResult['sites'], JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES);
+        $optionsJson = json_encode($sitesResult['sites'], JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES | JSON_HEX_TAG | JSON_HEX_AMP | JSON_HEX_APOS | JSON_HEX_QUOT);
 
         $body = <<<'HTML'
         <div class="p-card">